refactor: 설정 model 패키지 위치 변경 (component.model -> model)
