+++ android/app/src/main/java/com/yagubogu/ui/setting/SettingDeleteAccountScreen.kt
@@ -32,9 +32,9 @@ import com.google.firebase.analytics.analytics
 import com.yagubogu.R
 import com.yagubogu.ui.setting.component.SettingEventHandler
 import com.yagubogu.ui.setting.component.dialog.SettingDialog
-import com.yagubogu.ui.setting.component.model.MemberInfoItem
-import com.yagubogu.ui.setting.component.model.SettingDialogEvent
-import com.yagubogu.ui.setting.component.model.SettingEvent
+import com.yagubogu.ui.setting.model.MemberInfoItem
+import com.yagubogu.ui.setting.model.SettingDialogEvent
+import com.yagubogu.ui.setting.model.SettingEvent
 import com.yagubogu.ui.theme.Gray050
 import com.yagubogu.ui.theme.Gray100
 import com.yagubogu.ui.theme.Gray400
