+++ android/app/src/main/java/com/yagubogu/ui/setting/component/SettingEventHandler.kt
@@ -9,7 +9,7 @@ import com.yagubogu.R
 import com.yagubogu.presentation.login.LoginActivity
 import com.yagubogu.presentation.util.showToast
 import com.yagubogu.ui.main.MainActivity
-import com.yagubogu.ui.setting.component.model.SettingEvent
+import com.yagubogu.ui.setting.model.SettingEvent
 
 @Composable
 fun SettingEventHandler(
